@@ -1,19 +1,20 @@
-<?php declare(strict_types=1);
+<?php
 
 /**
  * It's free open-source software released under the MIT License.
  *
- * @author Anatoly Fenric <anatoly@fenric.ru>
- * @copyright Copyright (c) 2018, Anatoly Fenric
+ * @author Anatoly Nekhay <afenric@gmail.com>
+ * @copyright Copyright (c) 2018, Anatoly Nekhay
  * @license https://github.com/sunrise-php/http-client-curl/blob/master/LICENSE
  * @link https://github.com/sunrise-php/http-client-curl
  */
 
+declare(strict_types=1);
+
 namespace Sunrise\Http\Client\Curl;
 
-/**
- * Import classes
- */
+use CurlHandle;
+use CurlMultiHandle;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseFactoryInterface;
@@ -21,9 +22,6 @@ use Psr\Http\Message\ResponseInterface;
 use Sunrise\Http\Client\Curl\Exception\ClientException;
 use Sunrise\Http\Client\Curl\Exception\NetworkException;
 
-/**
- * Import functions
- */
 use function curl_close;
 use function curl_errno;
 use function curl_error;
@@ -33,21 +31,25 @@ use function curl_init;
 use function curl_multi_add_handle;
 use function curl_multi_close;
 use function curl_multi_exec;
+use function curl_multi_getcontent;
 use function curl_multi_init;
 use function curl_multi_remove_handle;
+use function curl_multi_select;
 use function curl_setopt_array;
 use function explode;
 use function in_array;
 use function ltrim;
+use function preg_split;
 use function sprintf;
+use function strpos;
 use function substr;
+use function usleep;
 
-/**
- * Import constants
- */
 use const CURLINFO_HEADER_SIZE;
 use const CURLINFO_RESPONSE_CODE;
 use const CURLINFO_TOTAL_TIME;
+use const CURLM_CALL_MULTI_PERFORM;
+use const CURLM_OK;
 use const CURLOPT_CUSTOMREQUEST;
 use const CURLOPT_HEADER;
 use const CURLOPT_HTTPHEADER;
@@ -55,121 +57,116 @@ use const CURLOPT_POSTFIELDS;
 use const CURLOPT_RETURNTRANSFER;
 use const CURLOPT_URL;
 
-/**
- * HTTP client based on cURL
- *
- * @link http://php.net/manual/en/intro.curl.php
- * @link https://curl.haxx.se/libcurl/c/libcurl-errors.html
- * @link https://www.php-fig.org/psr/psr-2/
- * @link https://www.php-fig.org/psr/psr-7/
- * @link https://www.php-fig.org/psr/psr-17/
- * @link https://www.php-fig.org/psr/psr-18/
- */
-class Client implements ClientInterface
+final class Client implements ClientInterface
 {
+    private const BODYLESS_HTTP_METHODS = ['HEAD', 'GET'];
+    private const DEFAULT_CURL_MULTI_SELECT_TIMEOUT = 1.0;
+    private const DEFAULT_CURL_MULTI_SELECT_SLEEP_DURATION = 1000;
+    private const REQUEST_TIME_HEADER_FIELD_NAME = 'X-Request-Time';
+    private const HEADER_FIELD_SEPARATOR = '/\r\n|\n/';
+
+    private ?CurlMultiHandle $curlMultiHandle = null;
 
     /**
-     * @var ResponseFactoryInterface
+     * @var array<array-key, CurlHandle>
      */
-    protected $responseFactory;
+    private array $curlHandles = [];
 
-    /**
-     * @var array<int, mixed>
-     */
-    protected $curlOptions;
-
-    /**
-     * Constructor of the class
-     *
-     * @param ResponseFactoryInterface $responseFactory
-     * @param array<int, mixed> $curlOptions
-     */
     public function __construct(
-        ResponseFactoryInterface $responseFactory,
-        array $curlOptions = []
+        private readonly ResponseFactoryInterface $responseFactory,
+        /** @var array<int, mixed> */
+        private readonly array $curlOptions = [],
+        private readonly ?float $curlMultiSelectTimeout = null,
+        private readonly ?int $curlMultiSelectSleepDuration = null,
     ) {
-        $this->responseFactory = $responseFactory;
-        $this->curlOptions = $curlOptions;
+    }
+
+    public function __destruct()
+    {
+        $this->clear();
     }
 
     /**
-     * {@inheritdoc}
+     * @inheritDoc
      */
-    public function sendRequest(RequestInterface $request) : ResponseInterface
+    public function sendRequest(RequestInterface $request): ResponseInterface
+    {
+        try {
+            return $this->executeRequest($request);
+        } finally {
+            $this->clear();
+        }
+    }
+
+    private function executeRequest(RequestInterface $request): ResponseInterface
+    {
+        return $request instanceof MultiRequest
+            ? $this->executeMultiRequest($request)
+            : $this->executeSingleRequest($request);
+    }
+
+    private function executeSingleRequest(RequestInterface $request): ResponseInterface
     {
         $curlHandle = $this->createCurlHandleFromRequest($request);
 
-        $isSuccess = curl_exec($curlHandle);
-        if ($isSuccess === false) {
-            throw new NetworkException($request, curl_error($curlHandle), curl_errno($curlHandle));
+        $curlExecuteResult = curl_exec($curlHandle);
+        if ($curlExecuteResult === false) {
+            throw new NetworkException(
+                $request,
+                curl_error($curlHandle),
+                curl_errno($curlHandle),
+            );
         }
 
-        $response = $this->createResponseFromCurlHandle($curlHandle);
-
-        curl_close($curlHandle);
-
-        return $response;
+        return $this->createResponseFromCurlHandle($curlHandle);
     }
 
-    /**
-     * Sends the given requests and returns responses in the same order
-     *
-     * @param RequestInterface ...$requests
-     *
-     * @return array<int, ResponseInterface>
-     *
-     * @throws ClientException
-     * @throws NetworkException
-     */
-    public function sendRequests(RequestInterface ...$requests) : array
+    private function executeMultiRequest(MultiRequest $multiRequest): MultiResponse
     {
-        /** @var list<RequestInterface> $requests */
+        $this->curlMultiHandle = curl_multi_init();
 
-        $curlMultiHandle = curl_multi_init();
-        if ($curlMultiHandle === false) {
-            throw new ClientException('Unable to create CurlMultiHandle');
+        foreach ($multiRequest->getRequests() as $key => $request) {
+            $curlHandle = $this->createCurlHandleFromRequest($request, $key);
+            $curlMultiStatusCode = curl_multi_add_handle($this->curlMultiHandle, $curlHandle);
+            ClientException::assertCurlMultiStatusCodeSame(CURLM_OK, $curlMultiStatusCode);
         }
 
-        $curlHandles = [];
-        foreach ($requests as $i => $request) {
-            $curlHandles[$i] = $this->createCurlHandleFromRequest($request);
-            curl_multi_add_handle($curlMultiHandle, $curlHandles[$i]);
-        }
+        $curlMultiSelectTimeout = $this->curlMultiSelectTimeout ?? self::DEFAULT_CURL_MULTI_SELECT_TIMEOUT;
+        // phpcs:ignore Generic.Files.LineLength.TooLong
+        $curlMultiSelectSleepDuration = $this->curlMultiSelectSleepDuration ?? self::DEFAULT_CURL_MULTI_SELECT_SLEEP_DURATION;
 
         do {
-            curl_multi_exec($curlMultiHandle, $isActive);
-        } while ($isActive);
+            $curlMultiStatusCode = curl_multi_exec($this->curlMultiHandle, $isCurlMultiExecuteStillRunning);
+            // https://stackoverflow.com/questions/19490837/curlm-call-multi-perform-deprecated
+            if ($curlMultiStatusCode === CURLM_CALL_MULTI_PERFORM) {
+                continue;
+            }
+
+            ClientException::assertCurlMultiStatusCodeSame(CURLM_OK, $curlMultiStatusCode);
+
+            if ($isCurlMultiExecuteStillRunning) {
+                $curlMultiSelectResult = curl_multi_select($this->curlMultiHandle, $curlMultiSelectTimeout);
+                if ($curlMultiSelectResult === -1) {
+                    // Take pauses to reduce CPU load...
+                    usleep($curlMultiSelectSleepDuration);
+                }
+            }
+        } while ($isCurlMultiExecuteStillRunning);
 
         $responses = [];
-        foreach ($curlHandles as $i => $curlHandle) {
-            $responses[$i] = $this->createResponseFromCurlHandle($curlHandle);
-            curl_multi_remove_handle($curlMultiHandle, $curlHandle);
-            curl_close($curlHandle);
+        foreach ($this->curlHandles as $key => $curlHandle) {
+            $responses[$key] = $this->createResponseFromCurlHandle($curlHandle);
         }
 
-        curl_multi_close($curlMultiHandle);
-
-        return $responses;
+        return new MultiResponse(...$responses);
     }
 
-    /**
-     * Creates a CurlHandle from the given request
-     *
-     * @param RequestInterface $request
-     *
-     * @return resource
-     *
-     * @throws ClientException
-     */
-    private function createCurlHandleFromRequest(RequestInterface $request)
+    private function createCurlHandleFromRequest(RequestInterface $request, int|string $key = 0): CurlHandle
     {
         $curlOptions = $this->curlOptions;
 
-        $curlOptions[CURLOPT_RETURNTRANSFER] = true;
-        $curlOptions[CURLOPT_HEADER] = true;
-
         $curlOptions[CURLOPT_CUSTOMREQUEST] = $request->getMethod();
-        $curlOptions[CURLOPT_URL] = $request->getUri()->__toString();
+        $curlOptions[CURLOPT_URL] = (string) $request->getUri();
 
         $curlOptions[CURLOPT_HTTPHEADER] = [];
         foreach ($request->getHeaders() as $name => $values) {
@@ -179,106 +176,101 @@ class Client implements ClientInterface
         }
 
         $curlOptions[CURLOPT_POSTFIELDS] = null;
-        if (!in_array($request->getMethod(), ['GET', 'HEAD'], true)) {
-            $curlOptions[CURLOPT_POSTFIELDS] = $request->getBody()->__toString();
+        if (!in_array($request->getMethod(), self::BODYLESS_HTTP_METHODS, true)) {
+            $curlOptions[CURLOPT_POSTFIELDS] = (string) $request->getBody();
         }
+
+        $curlOptions[CURLOPT_RETURNTRANSFER] = true;
+        $curlOptions[CURLOPT_HEADER] = true;
 
         $curlHandle = curl_init();
         if ($curlHandle === false) {
-            throw new ClientException('Unable to create CurlHandle');
+            throw new ClientException('Unable to create CurlHandle.');
         }
 
-        $isSuccess = curl_setopt_array($curlHandle, $curlOptions);
-        if ($isSuccess === false) {
-            throw new ClientException('Unable to configure CurlHandle');
+        $this->curlHandles[$key] = $curlHandle;
+
+        $curlSetOptionsResult = curl_setopt_array($curlHandle, $curlOptions);
+        if ($curlSetOptionsResult === false) {
+            throw new ClientException('Unable to configure CurlHandle.');
         }
 
         return $curlHandle;
     }
 
-    /**
-     * Creates a response from the given CurlHandle
-     *
-     * @param resource $curlHandle
-     *
-     * @return ResponseInterface
-     */
-    private function createResponseFromCurlHandle($curlHandle) : ResponseInterface
+    private function createResponseFromCurlHandle(CurlHandle $curlHandle): ResponseInterface
     {
-        /** @var int */
-        $statusCode = curl_getinfo($curlHandle, CURLINFO_RESPONSE_CODE);
-        $response = $this->responseFactory->createResponse($statusCode);
+        /** @var int $responseStatusCode */
+        $responseStatusCode = curl_getinfo($curlHandle, CURLINFO_RESPONSE_CODE);
+        $response = $this->responseFactory->createResponse($responseStatusCode);
 
-        /** @var float */
+        /** @var float $requestTime */
         $requestTime = curl_getinfo($curlHandle, CURLINFO_TOTAL_TIME);
-        $response = $response->withAddedHeader('X-Request-Time', sprintf('%.3f ms', $requestTime * 1000));
+        $formattedRequestTime = sprintf('%.3f ms', $requestTime * 1000);
+        $response = $response->withAddedHeader(self::REQUEST_TIME_HEADER_FIELD_NAME, $formattedRequestTime);
 
-        /** @var ?string */
-        $message = curl_multi_getcontent($curlHandle);
-        if ($message === null) {
-            return $response;
-        }
+        /** @var string $responseMessage */
+        $responseMessage = curl_multi_getcontent($curlHandle);
 
-        /** @var int */
-        $headerSize = curl_getinfo($curlHandle, CURLINFO_HEADER_SIZE);
-        $header = substr($message, 0, $headerSize);
-        $response = $this->populateResponseWithHeaderFields($response, $header);
+        /** @var int $responseHeaderSize */
+        $responseHeaderSize = curl_getinfo($curlHandle, CURLINFO_HEADER_SIZE);
+        $responseHeader = substr($responseMessage, 0, $responseHeaderSize);
+        $response = $this->populateResponseWithHeaderFields($response, $responseHeader);
 
-        $body = substr($message, $headerSize);
-        $response->getBody()->write($body);
+        $responseContent = substr($responseMessage, $responseHeaderSize);
+        $response->getBody()->write($responseContent);
 
         return $response;
     }
 
-    /**
-     * Populates the given response with the given header's fields
-     *
-     * @param ResponseInterface $response
-     * @param string $header
-     *
-     * @return ResponseInterface
-     *
-     * @link https://datatracker.ietf.org/doc/html/rfc7230#section-3.2
-     */
-    private function populateResponseWithHeaderFields(ResponseInterface $response, string $header) : ResponseInterface
+    private function populateResponseWithHeaderFields(ResponseInterface $response, string $header): ResponseInterface
     {
-        $fields = explode("\r\n", $header);
+        /** @var array<int, string> $fields */
+        // @phpstan-ignore varTag.nativeType
+        $fields = preg_split(self::HEADER_FIELD_SEPARATOR, $header);
 
         foreach ($fields as $i => $field) {
-            // The first line of a response message is the status-line, consisting
-            // of the protocol version, a space (SP), the status code, another
-            // space, a possibly empty textual phrase describing the status code,
-            // and ending with CRLF.
             // https://datatracker.ietf.org/doc/html/rfc7230#section-3.1.2
             if ($i === 0) {
                 continue;
             }
 
-            // All HTTP/1.1 messages consist of a start-line followed by a sequence
-            // of octets in a format similar to the Internet Message Format:
-            // zero or more header fields (collectively referred to as
-            // the "headers" or the "header section"), an empty line indicating the
-            // end of the header section, and an optional message body.
             // https://datatracker.ietf.org/doc/html/rfc7230#section-3
             // https://datatracker.ietf.org/doc/html/rfc5322
             if ($field === '') {
                 break;
             }
 
-            // While HTTP/1.x used the message start-line (see [RFC7230],
-            // Section 3.1) to convey the target URI, the method of the request, and
-            // the status code for the response, HTTP/2 uses special pseudo-header
-            // fields beginning with ':' character (ASCII 0x3a) for this purpose.
-            // https://datatracker.ietf.org/doc/html/rfc7540#section-8.1.2.1
-            if ($field[0] === ':') {
+            if (strpos($field, ':') === false) {
                 continue;
             }
 
-            [$name, $value] = explode(':', $field, 2);
+            /** @psalm-suppress PossiblyUndefinedArrayOffset */
+            [$fieldName, $fieldValue] = explode(':', $field, 2);
 
-            $response = $response->withAddedHeader($name, ltrim($value));
+            $response = $response->withAddedHeader($fieldName, ltrim($fieldValue));
         }
 
         return $response;
+    }
+
+    private function clear(): void
+    {
+        if ($this->curlMultiHandle instanceof CurlMultiHandle) {
+            foreach ($this->curlHandles as $curlSingleHandle) {
+                curl_multi_remove_handle($this->curlMultiHandle, $curlSingleHandle);
+            }
+        }
+
+        foreach ($this->curlHandles as $curlSingleHandle) {
+            curl_close($curlSingleHandle);
+        }
+
+        if ($this->curlMultiHandle instanceof CurlMultiHandle) {
+            curl_multi_close($this->curlMultiHandle);
+        }
+
+        $this->curlMultiHandle = null;
+        $this->curlHandles = [];
     }
 }
